+++ Hal/Heuristics/Cavity.h
@@ -4,12 +4,12 @@
 #include "../../Structure/Heuristic.h"
 #include "../../Utility/CLOCKWISE_FOR.hpp"
 
-class Cavity : public Heuristics<unsigned,Field>{
+class Cavity : public Heuristics<double,Field>{
 private:
 	void FillAround(Field& field,const Point& pos)const;
 
 public:
-	unsigned Execution(const Field& field);
+	double Execution(const Field& field);
 
 	Cavity();
 };
